Add custom assert macro
Fixes e.g. "control reaches end of non-void function" warnings from GCC and MSVC. The intention is to also add a proper way to provide a message string to the macro in the future.

Also move error.h/cpp to engine/ and fix "enumeration value 'None' not handled in switch" warning.